🚚 (model/xref) Extract canonical stuff to canonical model
